When folding param for loops, use int64_t for indices instead of int
This fixes a problem when the Chapel compiler is built with the Pathscale
compiler on XT in which a param for loop could cause the compiler to
incorrectly go into a very long loop.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15391 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ compiler/resolution/functionResolution.cpp
@@ -1927,13 +1927,13 @@ static void fold_param_for(CallExpr* loop) {
   block->insertAfter(noop);
   Symbol* index = toSymExpr(index_expr)->var;
   if (stride <= 0) {
-    for (int i = high; i >= low; i += stride) {
+    for (int64_t i = high; i >= low; i += stride) {
       SymbolMap map;
       map.put(index, new_IntSymbol(i));
       noop->insertBefore(block->copy(&map));
     }
   } else {
-    for (int i = low; i <= high; i += stride) {
+    for (int64_t i = low; i <= high; i += stride) {
       SymbolMap map;
       map.put(index, new_IntSymbol(i));
       noop->insertBefore(block->copy(&map));
